+++ tests/integration/test_regression.py
@@ -12,7 +12,7 @@ import torch
 from torch.utils.data import DataLoader
 
 from mol_gnn import models, nn
-from mol_gnn.nn import agg
+from mol_gnn.nn.gnn import agg
 from mol_gnn.nn.message_passing import edge, embed, message, update
 from mol_gnn.data import MoleculeDataset
